PYIC-5759: move vocab class helpers to test folder
