+++ libs/common-services/src/test/java/uk/gov/di/ipv/core/library/helpers/vocab/BirthDateGenerator.java
@@ -1,9 +1,9 @@
-package uk.gov.di.ipv.core.library.helpers;
+package uk.gov.di.ipv.core.library.helpers.vocab;
 
 import uk.gov.di.model.BirthDate;
 
-public class BirthDateHelper {
-    private BirthDateHelper() {}
+public class BirthDateGenerator {
+    private BirthDateGenerator() {}
 
     public static BirthDate createBirthDate(String value) {
         var birthDate = new BirthDate();
