+++ libs/common-services/src/test/java/uk/gov/di/ipv/core/library/helpers/vocab/NameGenerator.java
@@ -1,12 +1,12 @@
-package uk.gov.di.ipv.core.library.helpers;
+package uk.gov.di.ipv.core.library.helpers.vocab;
 
 import uk.gov.di.model.Name;
 import uk.gov.di.model.NamePart;
 
 import java.util.List;
 
-public class NameHelper {
-    private NameHelper() {}
+public class NameGenerator {
+    private NameGenerator() {}
 
     public static Name createName(List<NamePart> nameParts) {
         var name = new Name();
@@ -14,8 +14,8 @@ public class NameHelper {
         return name;
     }
 
-    public static class NamePartHelper {
-        private NamePartHelper() {}
+    public static class NamePartGenerator {
+        private NamePartGenerator() {}
 
         public static NamePart createNamePart(String value, NamePart.NamePartType type) {
             var namePart = new NamePart();
